+++ article_scraper/src/video_object.rs
@@ -87,6 +87,10 @@ impl VideoObject {
     }
 
     pub fn replace(&self, node: &mut Node) -> Result<(), FullTextParserError> {
+        if node.is_null() {
+            return Err(FullTextParserError::Xml);
+        }
+
         let mut parent = node.get_parent().ok_or(FullTextParserError::Xml)?;
         node.unlink();
